13-tensorflow-model/requirements.txt
tensorflow
tensorflow-text
tf-models-official
matplotlib
numpy